refactor: extract layout bridge module from monolithic lib.rs
Extracted 289 lines of shared layout computation into
crates/canvist_wasm/src/layout_bridge.rs:

- LayoutConstants (shared padding/style/config)
- ParagraphLayoutInfo (per-paragraph layout data)
- PARAGRAPH_SPACING constant
- layout_paragraphs() (full paragraph layout pipeline)
- find_para_and_line_for_offset() (offset → paragraph+line mapping)
- build_fragments() (styled run → TextFragment conversion)
- hit_x_on_line_ext() (x-position calculation within a line)

lib.rs: 12,388 → 12,104 lines (−284).
All pub(crate) — no public API changes.

177/177 tests pass, WASM builds with zero warnings.
